+++ app/models/additionals_import.rb
@@ -25,8 +25,6 @@ class AdditionalsImport < Import
       value = case v.custom_field.field_format
               when 'date'
                 row_date(row, "cf_#{v.custom_field.id}")
-              when 'list'
-                row_value(row, "cf_#{v.custom_field.id}").try(:split, ',')
               else
                 row_value(row, "cf_#{v.custom_field.id}")
               end
@@ -34,7 +32,7 @@ class AdditionalsImport < Import
 
       h[v.custom_field.id.to_s] =
         if value.is_a?(Array)
-          value.map { |val| v.custom_field.value_from_keyword(val.strip, object) }.flatten!.compact
+          value.map { |val| v.custom_field.value_from_keyword(val.strip, object) }.flatten!&.compact
         else
           v.custom_field.value_from_keyword(value, object)
         end