+++ app/overrides/wiki/show.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'wiki/show',
-                     name: 'addto-wiki-show',
-                     insert_before: 'div.contextual',
-                     original: '6b0cb1646d5e2cb23feee1805949e266036581e6',
-                     partial: 'wiki/show_additionals'