+++ lib/additionals/wiki_macros/last_updated_at_macro.rb
@@ -31,8 +31,9 @@ module Additionals
 
         return unless page
 
+        # TODO: find solution for time_tag without to use html_safe
         content_tag(:span,
-                    l(:label_updated_time, time_tag(page.updated_on)).html_safe,
+                    l(:label_updated_time, time_tag(page.updated_on)).html_safe, # rubocop:disable Rails/OutputSafety
                     class: 'last-updated-at')
       end
     end