+++ test/integration/routing_test.rb
@@ -8,4 +8,12 @@ class RoutingTest < Redmine::RoutingTest
   def test_issue_change_status
     should_route 'PUT /issues/1/change_status' => 'additionals_change_status#update', issue_id: '1'
   end
+
+  def test_help_macro
+    should_route 'GET /help/macros' => 'additionals_macros#show'
+  end
+
+  def test_auto_completes_fontawesome
+    should_route 'GET /auto_completes/fontawesome' => 'auto_completes#fontawesome'
+  end
 end